Drop javacc in favor of a manually written parser
MiniMessage language is too lenient for a strict parser like this to be
very good at handling. Rather than being invalid, any text which doesn't
match the correct format is simply ignored, which is hard to do with a
parser like this. A hand-written parser can have good performance to
accomplish this easier.